# Conflicts:
#	AbsServer/src/abs/constants/Constants.java
#	Engine/src/bank/logic/manager/BankManager.java

+++ Engine/src/bank/logic/loans/interest/Interest.java
@@ -10,7 +10,7 @@ public interface Interest extends Serializable {
 
     int getDuration();
 
-    float getPercent(); // returns interest's percent per payment
+    int getPercent(); // returns interest's percent per payment
     int getBaseAmount();
 
 }
